added validation class as separate file
