+++ 1-Ideal Weight/validation.rb
@@ -1,0 +1,28 @@
+class Validation
+  def initialize
+    @validations = []
+  end
+
+  def validate_int(number, name)
+    if number.to_i.to_s != number
+      @validations.push("#{name} must be integer!")
+    else
+      return number.to_i
+    end
+  end
+
+  def validate_presence_of(value, name)
+    if value.empty? == true
+      @validations.push("you must enter #{name}")
+    end
+  end
+
+  def check_all_validations
+    if @validations.empty? == false
+      @validations.each do |message|
+        puts message
+      end
+      exit 1
+    end
+  end
+end
